src/main/resources/faceRecognationServer/requirements.txt
deepface
opencv-python
mysql-connector-python
flask
numpy
pillow
tensorflow